メモがない場合のreturn処理を追加

+++ 04.class/actions/deleteMemo.js
@@ -2,6 +2,7 @@ import { selectMemo } from "../helpers/memoHelpers.js";
 
 export async function deleteMemo(database) {
   const memo = await selectMemo(database, "削除するメモを選んでください:");
+  if (!memo) return;
   await database.deleteMemoById(memo.id);
   console.log("メモを削除しました。");
 }
